+++ CHANGELOG.md
@@ -3,6 +3,7 @@ ChangeLog
 1.6.3 (2020-11-25)
 -------------------------
 - Add Logfile reload
+- Fix client sort order
 
 
 1.6.2 (2020-11-24)